requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
astroid==4.0.4
black==26.3.1
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
dill==0.4.1
fastapi==0.136.1
h11==0.16.0
idna==3.13
isort==8.0.1
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==26.2
pathspec==1.1.1
platformdirs==4.9.6
pydantic==2.13.4
pydantic_core==2.46.4
pylint==4.0.5
python-dotenv==1.2.2
python-multipart==0.0.27
pytokens==0.4.1
requests==2.33.1
starlette==1.0.0
tomlkit==0.14.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.7.0
uvicorn==0.46.0
